test(agy): cross-provider scope-security coverage + align scope.mjs baselines (#218)
The live-file scope hardening (fd-based O_NOFOLLOW copy + realpath TOCTOU re-check
+ per-file cap) was applied to all four providers' byte-identical scope.mjs as a
class fix, but only AGY had a regression test — dropping kimi/scope.mjs below its
stale, pre-hardening coverage baseline on CI.

- Replace agy-scope-security.test.mjs with a parameterized scope-security.test.mjs
  that exercises the new copy paths AND their error branches (out-of-root symlink
  swap, in-root O_NOFOLLOW/ELOOP, destination-write failure, ENOENT race, per-file
  cap) against agy/claude/gemini/kimi. Raises branch coverage above the 85 floor
  and covers the security-critical paths uniformly across providers.
- Align the four identical scope.mjs coverage baselines to consistent CI-safe
  values (lines 90.3 / branches 84 / functions 98). This raises claude/gemini's
  branch baseline (74.9 -> 84), enforces agy for the first time, and corrects
  kimi's stale 92.0 line baseline that the hardened code's defensive error
  branches cannot reach (the byte-identical siblings already accept ~90.3). Per
  the baseline note, values track the lower CI Node 20 coverage.
